－Add Some Demos－

+++ GJCFUitils/GJCFUitils/GJCFUitils/StringUitil/GJCFStringUitil.m
@@ -204,7 +204,7 @@
 {
     const char* aString = [string UTF8String];
     unsigned char result[16];
-    CC_MD5(aString, strlen(aString), result);
+    CC_MD5(aString, (unsigned int)strlen(aString), result);
     NSString* hash = [NSString stringWithFormat:@"%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X%02X",
                       result[0], result[1], result[2], result[3], result[4], result[5], result[6], result[7],
                       result[8], result[9], result[10], result[11], result[12], result[13], result[14], result[15]];
